Fix itemized schedule trigger errors
This changeset addresses an issue we are currently having with our itemized schedule triggers.  The triggers are setup to select rows from views currently, but they often run into a situation where the selection statement actually returns an empty resultset.  This in turn causes what amounts to an empty record (excluding what we ourselves add) being added to the itemized schedule queues to be processed later, only to fail because of trying to insert records that do not comply with table constraints (e.g., inserting a NULL value into a primary key column).

This puts us in a state of having to perpetually refresh data manually, which requires repartitioning the itemized schedule tables.  There is more work to be done to fix the root cause(s) of this problem, but this will at least prevent errors in our nightly refresh process for the time being.

I have also fixed a couple of issues with the integration tests in this.

+++ data/sql_setup/prepare_schedule_b.sql
@@ -32,25 +32,38 @@ declare
 begin
     if tg_op = 'INSERT' then
         select into view_row * from fec_vsum_sched_b_vw where sub_id = new.sub_id;
-        two_year_transaction_period_new = get_transaction_year(new.disb_dt, view_row.rpt_yr);
 
-        if two_year_transaction_period_new >= start_year then
-            delete from ofec_sched_b_queue_new where sub_id = view_row.sub_id;
-            insert into ofec_sched_b_queue_new values (view_row.*, timestamp, two_year_transaction_period_new);
+        -- Check to see if the resultset returned anything from the view.  If
+        -- it did not, skip the processing of the record, otherwise we'll end
+        -- up with a record full of NULL values.
+        -- "FOUND" is a PL/pgSQL boolean variable set to false initially in
+        -- any PL/pgSQL function and reset whenever certain statements are
+        -- run, e.g., a "SELECT INTO..." statement.  For more information,
+        -- visit here:
+        -- https://www.postgresql.org/docs/current/static/plpgsql-statements.html#PLPGSQL-STATEMENTS-DIAGNOSTICS
+        if found then
+            two_year_transaction_period_new = get_transaction_year(new.disb_dt, view_row.rpt_yr);
+
+            if two_year_transaction_period_new >= start_year then
+                delete from ofec_sched_b_queue_new where sub_id = view_row.sub_id;
+                insert into ofec_sched_b_queue_new values (view_row.*, timestamp, two_year_transaction_period_new);
+            end if;
+
+            return new;
         end if;
-
-        return new;
     elsif tg_op = 'UPDATE' then
         select into view_row * from fec_vsum_sched_b_vw where sub_id = new.sub_id;
 
-        two_year_transaction_period_new = get_transaction_year(new.disb_dt, view_row.rpt_yr);
+        if found then
+            two_year_transaction_period_new = get_transaction_year(new.disb_dt, view_row.rpt_yr);
 
-        if two_year_transaction_period_new >= start_year then
-            delete from ofec_sched_b_queue_new where sub_id = view_row.sub_id;
-            insert into ofec_sched_b_queue_new values (view_row.*, timestamp, two_year_transaction_period_new);
+            if two_year_transaction_period_new >= start_year then
+                delete from ofec_sched_b_queue_new where sub_id = view_row.sub_id;
+                insert into ofec_sched_b_queue_new values (view_row.*, timestamp, two_year_transaction_period_new);
+            end if;
+
+            return new;
         end if;
-
-        return new;
     end if;
 end
 $$ language plpgsql;
@@ -69,24 +82,38 @@ declare
 begin
     if tg_op = 'DELETE' then
         select into view_row * from fec_vsum_sched_b_vw where sub_id = old.sub_id;
-        two_year_transaction_period_old = get_transaction_year(view_row.disb_dt, view_row.rpt_yr);
 
-        if two_year_transaction_period_old >= start_year then
-            delete from ofec_sched_b_queue_old where sub_id = view_row.sub_id;
-            insert into ofec_sched_b_queue_old values (view_row.*, timestamp, two_year_transaction_period_old);
+        -- Check to see if the resultset returned anything from the view.  If
+        -- it did not, skip the processing of the record, otherwise we'll end
+        -- up with a record full of NULL values.
+        -- "FOUND" is a PL/pgSQL boolean variable set to false initially in
+        -- any PL/pgSQL function and reset whenever certain statements are
+        -- run, e.g., a "SELECT INTO..." statement.  For more information,
+        -- visit here:
+        -- https://www.postgresql.org/docs/current/static/plpgsql-statements.html#PLPGSQL-STATEMENTS-DIAGNOSTICS
+        if found then
+            two_year_transaction_period_old = get_transaction_year(view_row.disb_dt, view_row.rpt_yr);
+
+            if two_year_transaction_period_old >= start_year then
+                delete from ofec_sched_b_queue_old where sub_id = view_row.sub_id;
+                insert into ofec_sched_b_queue_old values (view_row.*, timestamp, two_year_transaction_period_old);
+            end if;
+
+            return old;
         end if;
-
-        return old;
     elsif tg_op = 'UPDATE' then
         select into view_row * from fec_vsum_sched_b_vw where sub_id = old.sub_id;
-        two_year_transaction_period_old = get_transaction_year(old.disb_dt, view_row.rpt_yr);
 
-        if two_year_transaction_period_old >= start_year then
-            delete from ofec_sched_b_queue_old where sub_id = view_row.sub_id;
-            insert into ofec_sched_b_queue_old values (view_row.*, timestamp, two_year_transaction_period_old);
+        if found then
+            two_year_transaction_period_old = get_transaction_year(old.disb_dt, view_row.rpt_yr);
+
+            if two_year_transaction_period_old >= start_year then
+                delete from ofec_sched_b_queue_old where sub_id = view_row.sub_id;
+                insert into ofec_sched_b_queue_old values (view_row.*, timestamp, two_year_transaction_period_old);
+            end if;
+
+            return new;
         end if;
-
-        return new;
     end if;
 end
 $$ language plpgsql;
